Extracting tests of database to a separated module
