+++ application/core/persistence_room/src/androidTest/java/io/writeopia/persistence/DocumentRepositoryTest.kt
@@ -4,11 +4,10 @@ import android.content.Context
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.writeopia.libraries.dbtests.DocumentRepositoryTests
 import io.writeopia.persistence.room.WriteopiaApplicationDatabase
 import io.writeopia.sdk.models.document.Document
 import io.writeopia.sdk.models.id.GenerateId
-import io.writeopia.sdk.models.story.StoryStep
-import io.writeopia.sdk.models.story.StoryTypes
 import io.writeopia.sdk.persistence.core.repository.DocumentRepository
 import io.writeopia.sdk.persistence.dao.DocumentEntityDao
 import io.writeopia.sdk.persistence.dao.StoryUnitEntityDao
@@ -19,7 +18,6 @@ import kotlinx.coroutines.test.runTest
 import kotlinx.datetime.Clock
 import org.junit.After
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -31,6 +29,8 @@ class DocumentRepositoryTest {
     private lateinit var documentEntityDao: DocumentEntityDao
     private lateinit var storyUnitEntityDao: StoryUnitEntityDao
     private lateinit var documentRepository: DocumentRepository
+
+    private lateinit var documentRepositoryTests: DocumentRepositoryTests
 
     @Before
     fun createDb() {
@@ -44,6 +44,7 @@ class DocumentRepositoryTest {
         storyUnitEntityDao = database.storyUnitDao()
 
         documentRepository = RoomDocumentRepository(documentEntityDao, storyUnitEntityDao)
+        documentRepositoryTests = DocumentRepositoryTests(documentRepository)
     }
 
     @After
@@ -73,128 +74,21 @@ class DocumentRepositoryTest {
 
     @Test
     fun saveSimpleDocumentInRepository() = runTest {
-        val id = GenerateId.generate()
-        val document = Document(
-            id = id,
-            title = "Document1",
-            content = emptyMap(),
-            createdAt = Clock.System.now(),
-            lastUpdatedAt = Clock.System.now(),
-            userId = "userId",
-        )
-
-        documentRepository.saveDocument(document)
-
-        val loadedDocument = documentEntityDao.loadDocumentById(id)
-        assertEquals(document, loadedDocument?.toModel())
+        documentRepositoryTests.saveAndLoadASimpleDocument()
     }
 
     @Test
     fun savingAndLoadingDocumentWithOneImageInRepository() = runTest {
-        val id = GenerateId.generate()
-        val document = Document(
-            id = id,
-            title = "Document1",
-            content = simpleImage(),
-            createdAt = Clock.System.now(),
-            lastUpdatedAt = Clock.System.now(),
-            userId = "userId",
-        )
-
-        documentRepository.saveDocument(document)
-        val loadedDocument = documentRepository.loadDocumentById(id)
-
-        assertEquals(document, loadedDocument)
+        documentRepositoryTests.savingAndLoadingDocumentWithOneImageInRepository()
     }
 
     @Test
     fun savingAndLoadingDocumentWithManyImagesInRepository() = runTest {
-        val id = GenerateId.generate()
-        val document = Document(
-            id = id,
-            title = "Document1",
-            content = imageStepsList(),
-            createdAt = Clock.System.now(),
-            lastUpdatedAt = Clock.System.now(),
-            userId = "userId",
-        )
-
-        documentRepository.saveDocument(document)
-        val loadedDocument = documentRepository.loadDocumentById(id)
-
-        assertTrue(storyUnitEntityDao.loadDocumentContent(id).isNotEmpty())
-        assertTrue(documentEntityDao.loadDocumentWithContentById(id)?.values?.isNotEmpty() ?: false)
-        assertEquals(document, loadedDocument)
+        documentRepositoryTests.savingAndLoadingDocumentWithManyImagesInRepository()
     }
-
 
     @Test
     fun savingAndLoadingDocumentOneImageGroupInRepository() = runTest {
-        val id = GenerateId.generate()
-        val document = Document(
-            id = id,
-            title = "Document1",
-            content = imageGroup(),
-            createdAt = Clock.System.now(),
-            lastUpdatedAt = Clock.System.now(),
-            userId = "userId",
-        )
-
-        documentRepository.saveDocument(document)
-        val loadedDocument = documentRepository.loadDocumentById(id)
-
-        assertEquals(document, loadedDocument)
+        documentRepositoryTests.savingAndLoadingDocumentOneImageGroupInRepository()
     }
-}
-
-fun simpleImage(): Map<Int, StoryStep> = mapOf(
-    0 to StoryStep(
-        localId = "1",
-        type = StoryTypes.IMAGE.type,
-    )
-)
-
-fun imageStepsList(): Map<Int, StoryStep> = mapOf(
-    0 to StoryStep(
-        localId = "1",
-        type = StoryTypes.IMAGE.type,
-    ),
-    1 to StoryStep(
-        localId = "2",
-        type = StoryTypes.IMAGE.type,
-    ),
-    2 to StoryStep(
-        localId = "3",
-        type = StoryTypes.IMAGE.type,
-    ),
-)
-
-fun imageGroup() : Map<Int, StoryStep>{
-    val groupId = GenerateId.generate()
-
-    return mapOf(
-        0 to StoryStep(
-            id = groupId,
-            localId = "1",
-            type = StoryTypes.GROUP_IMAGE.type,
-            steps = listOf(
-                StoryStep(
-                    localId = "2",
-                    type = StoryTypes.IMAGE.type,
-                    parentId = groupId,
-                ),
-                StoryStep(
-                    localId = "3",
-                    type = StoryTypes.IMAGE.type,
-                    parentId = groupId,
-                ),
-                StoryStep(
-                    localId = "4",
-                    type = StoryTypes.IMAGE.type,
-                    parentId = groupId,
-                )
-            )
-        ),
-    )
-
 }
