+++ libraries/dbtest/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    kotlin("multiplatform")
+}
+
+kotlin {
+    jvm {}
+
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(project(":writeopia"))
+                implementation(project(":writeopia_models"))
+                implementation(project(":plugins:writeopia_persistence_core"))
+                implementation(libs.kotlinx.datetime)
+                implementation(libs.kotlin.test)
+            }
+        }
+    }
+}
