+++ libraries/dbtest/src/commonMain/kotlin/io/writeopia/libraries/dbtests/DocumentRepositoryTests.kt
@@ -1,0 +1,135 @@
+package io.writeopia.libraries.dbtests
+
+import io.writeopia.sdk.models.document.Document
+import io.writeopia.sdk.models.id.GenerateId
+import io.writeopia.sdk.models.story.StoryStep
+import io.writeopia.sdk.models.story.StoryTypes
+import io.writeopia.sdk.persistence.core.repository.DocumentRepository
+import kotlinx.datetime.Clock
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
+
+class DocumentRepositoryTests(private val documentRepository: DocumentRepository) {
+
+    suspend fun saveAndLoadASimpleDocument() {
+        val id = GenerateId.generate()
+        val document = Document(
+            id = id,
+            title = "Document1",
+            content = emptyMap(),
+            createdAt = Clock.System.now(),
+            lastUpdatedAt = Clock.System.now(),
+            userId = "userId",
+        )
+
+        documentRepository.saveDocument(document)
+        val loadedDocument = documentRepository.loadDocumentById(id)
+
+        assertEquals(document, loadedDocument)
+    }
+
+    suspend fun savingAndLoadingDocumentWithOneImageInRepository() {
+        val id = GenerateId.generate()
+        val document = Document(
+            id = id,
+            title = "Document1",
+            content = simpleImage(),
+            createdAt = Clock.System.now(),
+            lastUpdatedAt = Clock.System.now(),
+            userId = "userId",
+        )
+
+        documentRepository.saveDocument(document)
+        val loadedDocument = documentRepository.loadDocumentById(id)
+
+        assertEquals(document, loadedDocument)
+    }
+
+    suspend fun savingAndLoadingDocumentWithManyImagesInRepository() {
+        val id = GenerateId.generate()
+        val document = Document(
+            id = id,
+            title = "Document1",
+            content = imageStepsList(),
+            createdAt = Clock.System.now(),
+            lastUpdatedAt = Clock.System.now(),
+            userId = "userId",
+        )
+
+        documentRepository.saveDocument(document)
+        val loadedDocument = documentRepository.loadDocumentById(id)
+
+        assertTrue(loadedDocument?.content?.isNotEmpty() == true)
+        assertEquals(document, loadedDocument)
+    }
+
+    suspend fun savingAndLoadingDocumentOneImageGroupInRepository() {
+        val id = GenerateId.generate()
+        val document = Document(
+            id = id,
+            title = "Document1",
+            content = imageGroup(),
+            createdAt = Clock.System.now(),
+            lastUpdatedAt = Clock.System.now(),
+            userId = "userId",
+        )
+
+        documentRepository.saveDocument(document)
+        val loadedDocument = documentRepository.loadDocumentById(id)
+
+        assertEquals(document, loadedDocument)
+    }
+
+}
+
+fun simpleImage(): Map<Int, StoryStep> = mapOf(
+    0 to StoryStep(
+        localId = "1",
+        type = StoryTypes.IMAGE.type,
+    )
+)
+
+fun imageStepsList(): Map<Int, StoryStep> = mapOf(
+    0 to StoryStep(
+        localId = "1",
+        type = StoryTypes.IMAGE.type,
+    ),
+    1 to StoryStep(
+        localId = "2",
+        type = StoryTypes.IMAGE.type,
+    ),
+    2 to StoryStep(
+        localId = "3",
+        type = StoryTypes.IMAGE.type,
+    ),
+)
+
+fun imageGroup() : Map<Int, StoryStep>{
+    val groupId = GenerateId.generate()
+
+    return mapOf(
+        0 to StoryStep(
+            id = groupId,
+            localId = "1",
+            type = StoryTypes.GROUP_IMAGE.type,
+            steps = listOf(
+                StoryStep(
+                    localId = "2",
+                    type = StoryTypes.IMAGE.type,
+                    parentId = groupId,
+                ),
+                StoryStep(
+                    localId = "3",
+                    type = StoryTypes.IMAGE.type,
+                    parentId = groupId,
+                ),
+                StoryStep(
+                    localId = "4",
+                    type = StoryTypes.IMAGE.type,
+                    parentId = groupId,
+                )
+            )
+        ),
+    )
+
+}
